# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API Terintegrasi menggunakan L5-Swagger",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server Environment"
        }
    ],
    "paths": {
...
        "/api/komplain": {
            "get": {
                "tags": [
                    "Data Komplain"
                ],
                "summary": "Get List Data Komplain (Paginated)",
                "description": "Mengambil semua data komplain bulan berjalan dengan filter pencarian, kategori keywords, status pengerjaan, dan tim PDE.",
                "operationId": "getDataKomplain",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan nama, pelapor, ruangan, permasalahan, atau nomor WA",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "printer"
                        }
                    },
                    {
                        "name": "kategori",
                        "in": "query",
                        "description": "Filter kategori berdasarkan kecocokan keyword masalah (simrs / maintanance)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "simrs",
                                "maintanance"
                            ]
                        }
                    },
                    {
                        "name": "is_done",
                        "in": "query",
                        "description": "Filter status pengerjaan (true: memiliki nomor_act, false: nomor_act kosong)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    },
                    {
                        "name": "recent",
                        "in": "query",
                        "description": "Jika is_done=false dan recent=true, tampilkan komplain 1 jam terakhir saja",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "nomor_act",
                        "in": "query",
                        "description": "Filter berdasarkan nomor_act penanganan tertentu, atau string 'null' untuk yang belum ditangani",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "null"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get data komplain success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "get data komplain success"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 12
                                                            },
                                                            "nama": {
                                                                "type": "string",
                                                                "example": "Roni"
                                                            },
                                                            "nama_pelapor": {
                                                                "type": "string",
                                                                "example": "Siti"
                                                            },
                                                            "ruangan": {
                                                                "type": "string",
                                                                "example": "Poli Jantung"
                                                            },
                                                            "permasalahan": {
                                                                "type": "string",
                                                                "example": "Printer macet tinta habis"
                                                            },
                                                            "nomor_wa": {
                                                                "type": "string",
                                                                "example": "08123456789"
                                                            },
                                                            "nomor_act": {
                                                                "type": "string",
                                                                "example": "085544332211",
                                                                "nullable": true
                                                            },
                                                            "tanggal": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2026-06-10"
                                                            },
                                                            "pde": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 2
                                                                    },
                                                                    "nama": {
                                                                        "type": "string",
                                                                        "example": "Ahmad PDE"
                                                                    },
                                                                    "alamat": {
                                                                        "type": "string",
                                                                        "example": "Gedung IT Lantai 2"
                                                                    }
                                                                },
                                                                "type": "object",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/komplain/dashboard": {
            "get": {
                "tags": [
                    "Data Komplain"
                ],
                "summary": "Get Counter Statisitik Dashboard",
                "description": "Mengambil rangkuman jumlah tiket open, done, pembagian kategori, beserta performa per tim PDE bulan berjalan.",
                "operationId": "getKomplainDashboard",
                "responses": {
                    "200": {
                        "description": "get data dashboard success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "get data dashboard success"
                                        },
                                        "data": {
                                            "properties": {
                                                "ticket_open": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "ticket_done": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "simrs_masuk": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "simrs_done": {
                                                    "type": "integer",
                                                    "example": 18
                                                },
                                                "maintenance_masuk": {
                                                    "type": "integer",
                                                    "example": 27
                                                },
                                                "maintenance_done": {
                                                    "type": "integer",
                                                    "example": 24
                                                },
                                                "pde_performance": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1,
                                                                "nullable": true
                                                            },
                                                            "nama": {
                                                                "type": "string",
                                                                "example": "Ahmad PDE"
                                                            },
                                                            "alamat": {
                                                                "type": "string",
                                                                "example": "Jember",
                                                                "nullable": true
                                                            },
                                                            "telp": {
                                                                "type": "string",
                                                                "example": "085544332211"
                                                            },
                                                            "total": {
                                                                "type": "integer",
                                                                "example": 15
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pde": {
            "get": {
                "tags": [
                    "Data Komplain"
                ],
                "summary": "Get Distinct Team PDE Data",
                "description": "Mengambil daftar unik tim PDE yang sudah pernah menangani komplain (berdasarkan unique nomor_act).",
                "operationId": "getTeamPde",
                "responses": {
                    "200": {
                        "description": "get data team PDE success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "get data team PDE success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1,
                                                        "nullable": true
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "PDE Team 1"
                                                    },
                                                    "alamat": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "telp": {
                                                        "type": "string",
                                                        "example": "085544332211"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Anda tanpa kata 'Bearer '",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}